gce: fix labels
label contained the envId as is, which does not adhere to the [allowed characters](https://cloud.google.com/compute/docs/labeling-resources#requirements).

fixed by storing a hash of the envId and profileId, in the labels, which are used as a filter when querying the instances. actual values are stored in the [instance metadata](https://cloud.google.com/compute/docs/metadata/setting-custom-metadata) and returned for the ls command.

fixes #153